tree: fix

+++ files/initrd/opt/arc/include/modules.sh
@@ -6,7 +6,7 @@ function getAllModules() {
   local PLATFORM=${1}
   local KVER=${2}
 
-  if [[ -z "${PLATFORM}" || -z "${KVER}" ]]; then
+  if [ -z "${PLATFORM}" ] || [ -z "${KVER}" ]; then
     echo ""
     return 1
   fi
@@ -41,8 +41,8 @@ function installModules() {
   shift 2
   local MLIST="${@}"
 
-  if [[ -z "${PLATFORM}" || -z "${KVER}" ]]; then
-    echo ""
+  if [ -z "${PLATFORM}" ] || [ -z "${KVER}" ]; then
+    echo "ERROR: installModules: Platform or Kernel Version not defined" >"${LOG_FILE}"
     return 1
   fi
   # Unzip modules for temporary folder
@@ -54,10 +54,14 @@ function installModules() {
   else
     tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" 2>"${LOG_FILE}"
   fi
-  local ODP="$(readConfigKey "arc.odp" "${USER_CONFIG_FILE}")"
+  if [ $? -ne 0 ]; then
+    return 1
+  fi
+
+  local ODP="$(readConfigKey "odp" "${USER_CONFIG_FILE}")"
   for F in $(ls "${TMP_PATH}/modules/"*.ko 2>/dev/null); do
     local M=$(basename ${F})
-    [ "${ODP}" = "true" ] && [ -f "${RAMDISK_PATH}/usr/lib/modules/${M}" ] && continue
+    [ "${ODP}" = "true" -a -f "${RAMDISK_PATH}/usr/lib/modules/${M}" ] && continue
     if echo "${MLIST}" | grep -wq "${M:0:-3}"; then
       cp -f "${F}" "${RAMDISK_PATH}/usr/lib/modules/${M}" 2>"${LOG_FILE}"
     else
@@ -70,8 +74,13 @@ function installModules() {
   else
     tar -zxf "${MODULES_PATH}/firmware.tgz" -C "${RAMDISK_PATH}/usr/lib/firmware" 2>"${LOG_FILE}"
   fi
+  if [ $? -ne 0 ]; then
+    return 1
+  fi
+
   # Clean
   rm -rf "${TMP_PATH}/modules"
+  return 0
 }
 
 ###############################################################################
@@ -84,16 +93,25 @@ function addToModules() {
   local KVER=${2}
   local KOFILE=${3}
 
-  if [[ -z "${PLATFORM}" || -z "${KVER}" || -z "${KOFILE}" ]]; then
+  if [ -z "${PLATFORM}" ] || [ -z "${KVER}" ] || [ -z "${KOFILE}" ]; then
     echo ""
     return 1
   fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
-  tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  local KERNEL="$(readConfigKey "arc.kernel" "${USER_CONFIG_FILE}")"
+  if [ "${KERNEL}" = "custom" ]; then
+    tar -zxf "${CKS_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  else
+    tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  fi
   cp -f ${KOFILE} ${TMP_PATH}/modules
-  tar -zcf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" .
+  if [ "${KERNEL}" = "custom" ]; then
+    tar -zcf "${CKS_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" .
+  else
+    tar -zcf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" .
+  fi
 }
 
 ###############################################################################
@@ -106,16 +124,25 @@ function delToModules() {
   local KVER=${2}
   local KONAME=${3}
 
-  if [[ -z "${PLATFORM}" || -z "${KVER}" || -z "${KOFILE}" ]]; then
+  if [ -z "${PLATFORM}" ] || [ -z "${KVER}" ] || [ -z "${KONAME}" ]; then
     echo ""
     return 1
   fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
-  tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  local KERNEL="$(readConfigKey "arc.kernel" "${USER_CONFIG_FILE}")"
+  if [ "${KERNEL}" = "custom" ]; then
+    tar -zxf "${CUSTOM_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  else
+    tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  fi
   rm -f ${TMP_PATH}/modules/${KONAME}
-  tar -zcf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" .
+  if [ "${KERNEL}" = "custom" ]; then
+    tar -zcf "${CUSTOM_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" .
+  else
+    tar -zcf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" .
+  fi
 }
 
 ###############################################################################
@@ -126,7 +153,7 @@ function delToModules() {
 function getdepends() {
   function _getdepends() {
     if [ -f "${TMP_PATH}/modules/${1}.ko" ]; then
-      depends=($(modinfo "${TMP_PATH}/modules/${1}.ko" | grep depends: | awk -F: '{print $2}' | awk '$1=$1' | sed 's/,/ /g'))
+      depends=($(modinfo "${TMP_PATH}/modules/${1}.ko" 2>/dev/null | grep depends: | awk -F: '{print $2}' | awk '$1=$1' | sed 's/,/ /g'))
       if [ ${#depends[@]} -gt 0 ]; then
         for k in ${depends[@]}; do
           echo "${k}"
@@ -139,15 +166,20 @@ function getdepends() {
   local KVER=${2}
   local KONAME=${3}
 
-  if [[ -z "${PLATFORM}" || -z "${KVER}" || -z "${KOFILE}" ]]; then
+  if [ -z "${PLATFORM}" ] || [ -z "${KVER}" ] || [ -z "${KONAME}" ]; then
     echo ""
     return 1
   fi
   # Unzip modules for temporary folder
   rm -rf "${TMP_PATH}/modules"
   mkdir -p "${TMP_PATH}/modules"
-  tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
-  DPS=($(_getdepends ${KONAME} | tr ' ' '\n' | sort -u))
+  local KERNEL="$(readConfigKey "arc.kernel" "${USER_CONFIG_FILE}")"
+  if [ "${KERNEL}" = "custom" ]; then
+    tar -zxf "${CUSTOM_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  else
+    tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+  fi
+  local DPS=($(_getdepends ${KONAME} | tr ' ' '\n' | sort -u))
   echo ${DPS[@]}
   rm -rf "${TMP_PATH}/modules"
 }
